+++ External/FEXCore/Source/Interface/Core/Interpreter/F80Ops.cpp
@@ -417,24 +417,6 @@ DEF_OP(F64SCALE) {
   memcpy(GDP, &Tmp, sizeof(double));
 }
 
-DEF_OP(VPCMPESTRX) {
-  const auto Op = IROp->C<IR::IROp_VPCMPESTRX>();
-  const auto Is64Bit = Op->GPRSize == 8;
-
-  const auto RAX = *GetSrc<uint64_t*>(Data->SSAData, Op->RAX);
-  const auto RDX = *GetSrc<uint64_t*>(Data->SSAData, Op->RDX);;
-  const auto LHS = *GetSrc<__uint128_t*>(Data->SSAData, Op->LHS);
-  const auto RHS = *GetSrc<__uint128_t*>(Data->SSAData, Op->RHS);
-
-  // We can be cheeky and encode the size at bit 8 to save a parameter
-  const auto Control = Op->Control | (uint16_t(Is64Bit) << 8);
-
-  const auto Result = OpHandlers<IR::OP_VPCMPESTRX>::handle(RAX, RDX, LHS, RHS, Control);
-
-  memset(GDP, 0, sizeof(uint64_t));
-  memcpy(GDP, &Result, sizeof(Result));
-}
-
 #undef DEF_OP
 
 } // namespace FEXCore::CPU
